RLSimion running on linux!!!! TODO: -debugging from VS there is some problem with the directory -release mode of all linux projects...shouldn't be too hard

+++ RLSimion/deep-vfa-policy.cpp
@@ -60,12 +60,12 @@ int DiscreteSoftmaxDeepPolicy::selectAction(const std::vector<double>& values)
 	//this is not cool, but we can safely overwrite the output values of the function
 	std::vector<double>& newValues = (std::vector<double>&) values;
 	double sum = 0;
-	for (int i = 0; i < newValues.size(); i++)
+	for (int i = 0; i < (int)newValues.size(); i++)
 	{
 		newValues[i] = std::exp(m_temperature->get() * newValues[i]);
 		sum += newValues[i];
 	}
-	for (int i = 0; i < newValues.size(); i++)
+	for (int i = 0; i < (int) newValues.size(); i++)
 	{
 		newValues[i] /= sum;
 	}
